Rename secure_channel/SecureChannelProtocol.h to secure_channel/Constants.h
